+++ ChangeLog.md
@@ -11,11 +11,18 @@ git branch -u origin/main main
 git remote set-head origin -a
 ```
 
+- *Geometry*
+  - Add a generic quick hull variant that can process arbitrary input lattice points and outputs their convex hull, even if it is not full dimensional (Jacques-Olivier Lachaud, [#1803](https://github.com/DGtal-team/DGtal/pull/1803))
+  - Add affine geometry module for determining the affine dimension of a subset of points or a set of vectors, and for providing bases and orthogonal bases in reduced-echelon form or in LLL-form (Jacques-Olivier Lachaud, [#1803](https://github.com/DGtal-team/DGtal/pull/1803))
+
 ## Changes
 
 - *IO*
   - Polyscope viewer upgraded from 2.4 to 2.5 (David Coeurjolly,  [#1796](https://github.com/DGtal-team/DGtal/pull/1796)). Reverting to 2.4  due to build issues (David Coeurjolly,  [#1797](https://github.com/DGtal-team/DGtal/pull/1797))
   - The DGTAL_BUILD_POLYSCOPE_EXAMPLES option has been removed. If the polyscope viewer is enabled (DGTAL_WITH_POLYSCOPE_VIEWER), the polyscope examples are automatically built (David Coeurjolly, [#1814](https://github.com/DGtal-team/DGtal/pull/1814))
+
+- *Geometry*
+  - ConvexityHelper::computeLatticePolytope can now outputs exact lattice polytopes for k-dimensional sets of points in dD, whenever k in 0,1,d-1,d (Jacques-Olivier Lachaud, [#1803](https://github.com/DGtal-team/DGtal/pull/1803))
 
 - *Documentation*
   - Refactoring of the volumetric examples in the Shortcuts (David Coeurjolly, [#1807](https://github.com/DGtal-team/DGtal/pull/1807))
@@ -28,6 +35,7 @@ git remote set-head origin -a
 
 - *Geometry*
   - Fix issue [#1797](https://github.com/DGtal-team/DGtal/pull/1797), assertions raised in testCellGeometry and testDigitalConvexity because of improper digital space extent (Jacques-Olivier Lachaud, [#1801](https://github.com/DGtal-team/DGtal/pull/1801))
+  - QuickHull could sometimes output that its output was not full dimensional since the computation of affine dimension was randomized. Now QuickHull relies on an exact computation of affine dimension (Jacques-Olivier Lachaud, [#1803](https://github.com/DGtal-team/DGtal/pull/1803))
   - Silenting a warning in the Naive3DDSSComputer.ih file. (David Coeurjolly, [#1815](https://github.com/DGtal-team/DGtal/pull/1815))
 
 - *Build*